+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## Unreleased
+
+### Added
+
+- Optional request body parameter `idempotenceKey` for `::userReportMake` in `ClientInterface` and `Client`
+
 ## v3.4.0
 
 ### Changed